We changed some of the original code and added our own
We changed some of the original code for the ingrediants in the ingrediants.css, changing width height and color. We also added our own ingrediant to the ingrediants.css code page (we added the bacon). On the index.html page we moved around some of the ingrediants and added ours.